Protect mutation endpoints from unauthorized users.

+++ src/resolvers/Mutation.js
@@ -1,6 +1,8 @@
 import pubsub from '../pubsub.js';
 
 const createCar = (parent, args, context, info) => {
+	if (!context.userId)
+		throw new Error('You have no access to reach this endpoint!');
 	const myCar = { ...args };
 	pubsub.publish('CAR_CREATED', myCar);
 	context.carController.insertCar(myCar);
@@ -8,6 +10,8 @@ const createCar = (parent, args, context, info) => {
 };
 
 const deleteCar = (parent, args, context, info) => {
+	if (!context.userId)
+		throw new Error('You have no access to reach this endpoint!');
 	context.carController.deleteCarById(args.id);
 	return 'Deleted.';
 };
